umage-grayscale/requirements.txt
pillow>=10.0.0,<11.0.0
